Changed Card content and Python Scripts
